Memory optimization for Render free tier 512MB limit
- StackSTAC: float64 → float32 (halves raster memory, sufficient for spectral indices)
- Added gc.collect() after index computation and before change detection
- Prevents OOM during full analysis pipeline on Render free tier

+++ analysis-service/app/services/stackstac_adapter.py
@@ -252,7 +252,7 @@ def stackstac_compute_index(
         resolution=resolution,
         epsg=epsg,
         max_dim=max_dim,
-        dtype="float64",
+        dtype="float32",  # float32 saves 50% memory vs float64 — sufficient precision for spectral indices
     )
 
     data = result["data"]  # (n_bands, height, width)
